Big file movement for Ember 2 compatibility - unnecessary components removed
