+++ addon/components/form-submission-button.js
@@ -35,4 +35,5 @@ export default Ember.Component.extend({
 
     this.sendAction();
   },
+
 });
