+++ addon/components/form-submission.js
@@ -2,8 +2,6 @@ import Ember from 'ember';
 import FormSubmissionClassNameMixin from 'ember-easy-form-extensions/mixins/components/form-submission-class-name';
 import layout from '../templates/components/form-submission';
 
-const { computed } = Ember;
-
 export default Ember.Component.extend(
   FormSubmissionClassNameMixin, {
 
@@ -25,7 +23,7 @@ export default Ember.Component.extend(
 
   /* Properties */
 
-  classNameBindings: ['className'],
+  classNameBindings: ['className', 'submissionClassName'],
   formIsSubmitted: false,
   layout: layout,
 
@@ -46,6 +44,21 @@ export default Ember.Component.extend(
     },
 
   },
+
+  /* CPs */
+
+  submissionClassName: computed('className', 'formIsSubmitted',
+    function() {
+      const className = this.get('className');
+      const formIsSubmitted = this.get('formIsSubmitted');
+
+      if (formIsSubmitted) {
+        return `${className}-submitted`;
+      } else {
+        return null;
+      }
+    }
+  ),
 
 
 });
