--- addon/mixins/components/form-submission-class-name.js
@@ -1,21 +1,0 @@
-import Ember from 'ember';
-
-const { computed } = Ember;
-
-export default Ember.Mixin.create({
-  className: null,
-  classNameBindings: ['submissionClassName'],
-
-  submissionClassName: computed('className', 'formIsSubmitted',
-    function() {
-      const className = this.get('className');
-      const formIsSubmitted = this.get('formIsSubmitted');
-
-      if (formIsSubmitted) {
-        return `${className}-submitted`;
-      } else {
-        return null;
-      }
-    }
-  ),
-});
