--- addon/utils/default-for.js
@@ -1,7 +1,0 @@
-export default function defaultFor(variable, defaultValue) {
-  if (typeof variable !== 'undefined' && variable !== null) {
-    return variable;
-  } else {
-    return defaultValue;
-  }
-}
