+++ bower.json
@@ -1,18 +1,9 @@
 {
   "name": "ember-easy-form-extensions",
   "dependencies": {
-    "ember": "2.3.1",
+    "ember": "~2.3.1",
     "ember-cli-shims": "0.1.0",
-    "ember-cli-test-loader": "0.2.1",
-    "ember-load-initializers": "0.1.7",
-    "ember-qunit": "0.4.16",
-    "ember-qunit-notifications": "0.1.0",
-    "ember-resolver": "~0.1.20",
-    "jquery": "1.11.3",
-    "loader.js": "ember-cli/loader.js#3.4.0",
-    "qunit": "~1.20.0"
-  },
-  "resolutions": {
-    "ember": "2.3.1"
+    "ember-cli-test-loader": "0.2.2",
+    "ember-qunit-notifications": "0.1.0"
   }
 }
