+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ember-easy-form-extensions",
-  "version": "1.0.2",
-  "description": "Enhances Ember EasyForm by providing easy action handling, validations, and Ember 1.13 support for your forms",
+  "version": "2.0.0",
+  "description": "Enhances Ember form addons and components, like Easy Form, with Boilerplate action handling, validations, and more",
   "directories": {
     "doc": "doc",
     "test": "tests"
@@ -19,33 +19,31 @@
   "license": "MIT",
   "devDependencies": {
     "broccoli-asset-rev": "^2.2.0",
-    "ember-cli": "1.13.15",
+    "ember-ajax": "0.7.1",
+    "ember-cli": "2.3.0",
     "ember-cli-app-version": "^1.0.0",
-    "ember-cli-content-security-policy": "0.4.0",
-    "ember-cli-dependency-checker": "^1.1.0",
+    "ember-cli-dependency-checker": "^1.2.0",
     "ember-cli-htmlbars-inline-precompile": "^0.3.1",
-    "ember-cli-ic-ajax": "0.2.4",
     "ember-cli-inject-live-reload": "^1.3.1",
-    "ember-cli-moment-shim": "1.0.0",
-    "ember-cli-qunit": "^1.0.4",
+    "ember-cli-qunit": "^1.2.1",
     "ember-cli-release": "0.2.8",
-    "ember-cli-sri": "^1.2.0",
+    "ember-cli-sri": "^2.0.0",
     "ember-cli-uglify": "^1.2.0",
-    "ember-data": "2.3.3",
-    "ember-disable-prototype-extensions": "^1.0.0",
+    "ember-data": "^2.3.0",
+    "ember-disable-prototype-extensions": "^1.1.0",
     "ember-disable-proxy-controllers": "^1.0.1",
     "ember-export-application-global": "^1.0.4",
-    "ember-moment": "6.0.0",
-    "ember-try": "~0.0.8",
-    "moment": "2.11.2",
-    "moment-timezone": "0.5.0",
-    "ember-easy-form": "1.0.0-rc.1"
+    "ember-load-initializers": "^0.5.0",
+    "ember-resolver": "^2.0.3",
+    "ember-try": "^0.1.2",
+    "loader.js": "^4.0.0"
   },
   "keywords": [
     "ember-addon",
     "ember",
     "easy-form",
     "forms",
+    "form",
     "input",
     "submission",
     "save",
@@ -57,8 +55,8 @@
   ],
   "dependencies": {
     "ember-cli-babel": "^5.1.5",
-    "ember-validations": "2.0.0-alpha.4",
-    "ember-cli-htmlbars": "^1.0.1"
+    "ember-cli-htmlbars": "^1.0.1",
+    "ember-validations": "2.0.0-alpha.4"
   },
   "ember-addon": {
     "configPath": "tests/dummy/config"
