--- tests/dummy/app/resources/application/template.hbs
@@ -1,12 +1,0 @@
-<h2 id="title">ember-easy-form-extensions</h2>
-<dl>
-  <dt>Current path</dt>
-  <dd>{{currentPath}}</dd>
-</dl>
-
-<ul>
-  <li>{{#link-to 'fruit.form'}}Fruit Form{{/link-to}}</li>
-  <li>{{#link-to 'fruit.model-path'}}Fruit Model Path{{/link-to}}</li>
-</ul>
-
-{{outlet}}
