--- tests/dummy/app/resources/fruit/form/template.hbs
@@ -1,25 +1,0 @@
-{{#form-for model}}
-  {{input-for 'name'}}
-  {{input-for 'description' as='text'}}
-  {{!-- {{input-for 'color' as='select'}} --}}
-  {{input-for 'numberOfSeeds' as='number'}}
-  {{input-for 'isTropical' as='checkbox'}}
-  {{!-- {{input-for 'pickedOn' as='date'}} --}}
-
-  {{form-submission}}
-{{/form-for}}
-
-<dl>
-  <dt>Name</dt>
-  <dd data-test="model-name">{{model.name}}</dd>
-  <dt>Description</dt>
-  <dd data-test="model-description">{{model.description}}</dd>
-  {{!-- <dt>Category</dt>
-  <dd data-test="model-color">{{model.color}}</dd> --}}
-  <dt>Number of Seeds</dt>
-  <dd data-test="model-number-of-seeds">{{model.numberOfSeeds}}</dd>
-  <dt>Is Tropical</dt>
-  <dd data-test="model-is-tropical">{{model.isTropical}}</dd>
-  {{!-- <dt>Picked On</dt>
-  <dd data-test="model-picked-on">{{model.pickedOn}}</dd> --}}
-</dl>
