--- tests/dummy/app/resources/fruit/model-path/controller.js
@@ -1,13 +1,0 @@
-import Ember from 'ember';
-import FormMixin from 'ember-easy-form-extensions/mixins/controllers/form';
-
-export default Ember.Controller.extend(
-  FormMixin, {
-
-  model: {
-    fruit: {
-      name: null
-    }
-  },
-
-});
