--- tests/dummy/app/sdf/application.js
@@ -1,31 +1,0 @@
-import DS from 'ember-data';
-import FixtureAdapter from 'ember-data-fixture-adapter';
-
-export default FixtureAdapter.extend({
-  simulateRemoteResponse: false,
-
-  /**
-  By default, Ember Data does not implement a method with which to query features. Thus, we have to add it ourselves.
-
-  @todo Return single object if only one record is found
-  */
-
-  queryFixtures: function(records, query /*, type */) {
-    return records.filter(function(record) {
-      for (var key in query) {
-        var value;
-
-        if (!query.hasOwnProperty(key)) {
-          continue;
-        }
-
-        value = query[key];
-
-        if (record[key] !== value) {
-          return false;
-        }
-      }
-      return true;
-    });
-  }
-});
