--- tests/dummy/app/serializers/application.js
@@ -1,3 +1,0 @@
-import DS from 'ember-data';
-
-export default DS.RESTSerializer.extend();
