+++ tests/dummy/config/environment.js
@@ -3,7 +3,6 @@
 module.exports = function(environment) {
   var ENV = {
     modulePrefix: 'dummy',
-    podModulePrefix: 'dummy/resources',
     environment: environment,
     baseURL: '/',
     locationType: 'auto',
@@ -17,10 +16,6 @@ module.exports = function(environment) {
     APP: {
       // Here you can pass flags/options to your application instance
       // when it is created
-    },
-
-    contentSecurityPolicy: {
-      'style-src': "'self' 'unsafe-inline'"
     }
   };
