+++ tests/integration/components/form-submission-button-test.js
@@ -1,0 +1,24 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('form-submission-button', 'Integration | Component | form submission button', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });"
+
+  this.render(hbs`{{form-submission-button}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:"
+  this.render(hbs`
+    {{#form-submission-button}}
+      template block text
+    {{/form-submission-button}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
